Adds link tag back to <head> for PNG favicon.
Chrome and Safari will pick the .ico if there are links to both [according to this blog post](http://www.jonathantneal.com/blog/understand-the-favicon/).

The recommended approach from the blog post if you want to serve PNG favicons to browsers that support them is to just have a `<link>` for the PNG favicon. This will get ignored by browsers that don't support PNGs (IE10 and below) and they will then load the default `/favicon.ico`.

This way we can have PNGs in Chrome and Safari and ICO for fallbacks.