Extract release posts to collection (#214)
This change makes use of Jekyll's collection feature to separate release notes as a distinct content type. We basically did this already in treating release notes and blog posts differently in the template. Thus this change has no visible effect. The generated result is identical. It's just an internal refactoring for better house keeping.

Practical benefit: We no longer need to update the latest_version config data, it's automatically extracted from the latest entry in the releases collection.