Remove broken links to files from ancient commits
I originally tried to fix these links in #89 by replacing the
non-permalink `master` links to a commit that corresponds to the time of
publication. But apparently that didn't always work. For some reason
it appears these files didn't even exist on master when they were
mentioned in blog posts. Not sure what's going on there, but the links
are not that relevant and it doesn't make much sense to dig deeper into
this.