Fix updating `all_time` value for *other* sponsors (#854)
For sponsors who are not tracked through OpenCollective we need to update the `all_time` value explicitly.
#710 introduced an automation for this but it was completely flawed and never actually worked as intended (see https://github.com/crystal-lang/crystal-website/pull/851#discussion_r1791570578).

With this patch we keep track of `time_last_payment` in `_data/others.json`. The original value is set to March 2024 when we last updated the current `all_time` values.
The merge script now automatically adds the amount of the current montly contribution for every month since the last update. Usually the script runs at least once per month, so it should only add one amount. But it also catches up gracefully if there's an interruption (and it works for the initial catch up).
The script then writes the updated values for `all_time` and `time_last_payment` (set to now) back to `_data/others.json` and commits it so we keep track of which monthly contribution has already been processed.