Add separate visual category for Gold Sponsors (#857)
Previously, Gold and Corporate Sponsors were mixed into the same group for logo placement on the homepage.
We noticed that this is not an accurate representation considering the the importance of our major sponsors.
So this introduces a separate category for Gold Sponsors, with a logo size between the Silver and Corporate sponsors. In the process, we're adjusting the existing sizes a bit to make it fit.