Add links to external discussions (#914)
When there are good discussions about blog posts (or releases etc.) elsewhere on the web, it's great to link there.
This patch does that for the current 1.16.0 release notes, which are discussed on Hacker News.

Idea inspired by https://jonathan-frere.com/posts/adding-discussions/
The process of adding backlinks could even be automated in the future.