[CI] Add pre-commit lint workflow (#926)
Runs pre-commit from devenv in CI to enforce style rules are applied on each change.
This workflow only considers the files that are touched by respective PR. It does not run full checks on the entire code base.
It's essentially the CI-version of that local pre-commit hook.

Alternatively, we could run pre-commit with `--all-files` in order to check the entire code base. This might be better in the long run to avoid missing some violations.
For now, it might be best to only consider the affected files, while we need to bring legacy code up to speed (#927 for example).